fix(ci): satisfy catalog and file shape checks
