+++ catalogs/agents/registry.json
@@ -1,6 +1,6 @@
 {
   "schemaVersion": 1,
-  "registryVersion": "2026-07-14.1",
+  "registryVersion": "2026-07-15.1",
   "generatedAt": "2026-05-31T00:00:00Z",
   "agents": [
     {
